Allow enabling http request/response debug
Creating a client as usual will work as before:

```
client := morpheus.NewClient(c.Url)
```

but we add the ability to create a client with debug set
to true with:

```
client := morpheus.NewClient(c.Url, morpheus.WithDebug(true))
```

this will log http requests and responses to stderr.

We add a note to point out that if debug is enabled this will
log all http request/response contents, so should be
used with caution.

Example snippet of output:

```
  | \tX-Content-Type-Options: nosniff
  | \tX-Frame-Options: SAMEORIGIN
  | BODY         :
  | {
  |    "provisioningSettings": {
  |       "allowZoneSelection": false,
  |       "allowServerSelection": false,
.
.
.
```